chore: use action instead of api for creation
